# perf: Avoid unnecessary URL processing while parsing links