--- a/news/13132.feature.rst
+++ b/news/13132.feature.rst
@@ -1,0 +1,1 @@
+Optimize package collection by avoiding unnecessary URL parsing and other processing.