@ichard26 2024-12-27
As an example for where this matters, this saves about 1600 ms while collecting and resolving a list of homeassistant dependencies:

Command: `python -m cProfile -o profile2.pstats -m pip install -r temp/homeassistant/requirements.txt --dry-run`

<details><summary>Profile (before)</summary>
<p>
![image](https://github.com/user-attachments/assets/b1e5771a-c9b5-47e4-8f98-6badaaebbe72)
</p>
</details> 

<details><summary>Profile (after)</summary>
<p>
![image](https://github.com/user-attachments/assets/7597f4b3-ec37-4f6a-893b-87fb07c4e917)
</p>
</details> 

| Method | Before | After |
|--------|--------|--------|
| `Link.from_json` | 3310 ms | 1730 ms |
|  `LinkEvaluator.evaluate_link` | 1040 ms | 990 ms |

And while this depends on network performance (so please look at the time elapsed, not the percentages which may be off), the entire command takes ~16-18 seconds, so the savings are significant.