@ichard26 2025-01-05
Thanks for the review @sbidoul. To be safe, I threw the optimized `_absolute_link_url()` function at the stdlib urllib.parse test suite and it passed.